+++ src/ChoiceSubject/ChoiceSubjectElement.mjs
@@ -88,10 +88,7 @@ export class ChoiceSubjectElement extends HTMLElement {
             mode: "closed"
         });
 
-        flux_css_api.adopt(
-            this.#shadow,
-            css
-        );
+        this.#shadow.adoptedStyleSheets.push(css);
 
         this.#render();
     }