+++ src/Create/CreateElement.mjs
@@ -81,10 +81,7 @@ export class CreateElement extends HTMLElement {
             mode: "closed"
         });
 
-        flux_css_api.adopt(
-            this.#shadow,
-            css
-        );
+        this.#shadow.adoptedStyleSheets.push(css);
 
         this.#render();
     }