+++ src/FormButton/FormButtonElement.mjs
@@ -41,10 +41,7 @@ export class FormButtonElement extends HTMLElement {
             mode: "closed"
         });
 
-        flux_css_api.adopt(
-            this.#shadow,
-            css
-        );
+        this.#shadow.adoptedStyleSheets.push(css);
 
         this.#render();
     }