+++ src/IdentificationNumber/IdentificationNumberElement.mjs
@@ -83,10 +83,7 @@ export class IdentificationNumberElement extends HTMLElement {
             mode: "closed"
         });
 
-        flux_css_api.adopt(
-            this.#shadow,
-            css
-        );
+        this.#shadow.adoptedStyleSheets.push(css);
 
         this.#render();
     }