+++ src/IntendedDegreeProgram/IntendedDegreeProgramElement.mjs
@@ -92,10 +92,7 @@ export class IntendedDegreeProgramElement extends HTMLElement {
             mode: "closed"
         });
 
-        flux_css_api.adopt(
-            this.#shadow,
-            css
-        );
+        this.#shadow.adoptedStyleSheets.push(css);
 
         this.#render();
     }