+++ src/Mandatory/MandatoryElement.mjs
@@ -39,10 +39,7 @@ export class MandatoryElement extends HTMLElement {
             mode: "closed"
         });
 
-        flux_css_api.adopt(
-            this.#shadow,
-            css
-        );
+        this.#shadow.adoptedStyleSheets.push(css);
 
         this.#render();
     }