+++ src/PersonalData/PersonalDataElement.mjs
@@ -106,10 +106,7 @@ export class PersonalDataElement extends HTMLElement {
             mode: "closed"
         });
 
-        flux_css_api.adopt(
-            this.#shadow,
-            css
-        );
+        this.#shadow.adoptedStyleSheets.push(css);
 
         this.#render();
     }