+++ src/Portrait/PortraitElement.mjs
@@ -102,10 +102,7 @@ export class PortraitElement extends HTMLElement {
             mode: "closed"
         });
 
-        flux_css_api.adopt(
-            this.#shadow,
-            css
-        );
+        this.#shadow.adoptedStyleSheets.push(css);
 
         this.#render();
     }