+++ src/UniversityEntranceQualification/UniversityEntranceQualificationElement.mjs
@@ -90,10 +90,7 @@ export class UniversityEntranceQualificationElement extends HTMLElement {
             mode: "closed"
         });
 
-        flux_css_api.adopt(
-            this.#shadow,
-            css
-        );
+        this.#shadow.adoptedStyleSheets.push(css);
 
         this.#render();
     }